Fix wrong version being reported
Cinderlib CSI was reporting the wrong version on startup, as it was
hardcoded in the `cinderlib_csi.py` file and not being bumped.

Instead of importing `cinderlib_csi/__init__.py` file from
`cinderlib_csi.py`, which doesn't look too clean, we add that location
to the bumpversion configuration file so it gets automaticaly bumped.