Change readme to Markdown
Docker Hub doesn't know how to render ReStructuredText, so we end up
with a mix of rendered and unrendered readme on automatic builds.

Changing README.rst to README.md and changing contents to Markdown so it
can be properly rendered.

The rendered used by Docker Hub is not as nice as the one from GitHub,
so we are adapting the Markdown so it doesn't look as bad on Docker
builds.