On node publish check the staging
When receiving a node publish request don't assume that just because the
file/directory exists it means that it has what we expect it to have.

Verify that the private_bind file is mount binded to a device.  If it's
a block device, verify that staging points to the private bind.  If it's
a "mount", check that the staging directory is properly mounted.