Fix long description on PyPi
When we converted the readme from RST to Markdown we broke the auto
filling of the long description on setup.py and the rendering of it on
PyPi.

We already fixed the loading to the long description, now we fix the
rendering.

To fix the rendering we define the content type on setup.py and use
twine to update.

We also convert HISTORY from RST to Markdown, because it is collated
with the README to generate the long description.

While we are at it we also introduce an horizontal rule to separate the
readme from the history on PyPi's long description.

Dev requirements have been updated with the specific setuptools, wheel,
and twine versions required to support Markdown.