Fix block stage and publish
The patch checking for block failed on a second call to NodeStageVolume
if the first one was successful, and also on the call to
NodePublishVolume.

The cause is that we were checking if the staging path was a regular
file, which is right on the first call of staging, but false on the
second when we have already bind mounted it, and is false on the publish
call for the staging path since we've already bind mounted it.

Fix _check_path method to accept regular and block file on paths.