Support controller unpublish with no node_id
Trying to unpublish all connections from a volume on the controller
would return a "list index out of range" error.

The CSI spec says that the controller should unpublish all existing
connections on the volume when the node_id is not present.

Support this behavior and disconnect all existing connections on the
volume if node_id is not present.