Support live debugging
To facilitate debugging of the plugin we now support live debugging with
pdb and rpdb.

We have to enable it using the X_CSI_DEBUG_MODE environmental variable,
and then turn it ON/OFF with the SIGUSR1 signal.

We include our own rpdb file since the one from PyPi breaks with
OS-Brick.