Improve node staging of mount
When doing a node staging of a mount type of volume with multipathing we
may end up seeing an error such as:

 Traceback (most recent call last):
   File "csi/cinderlib_csi.py", line 118, in dolog
     result = f(self, request, context)
   File "csi/cinderlib_csi.py", line 155, in checker
     return f(self, request, context)
   File "csi/cinderlib_csi.py", line 800, in NodeStageVolume
     context)
   File "csi/cinderlib_csi.py", line 708, in _format_device
     stdout, stderr = self.sudo('lsblk', '-nlfoFSTYPE', device, retries=4)
   File "csi/cinderlib_csi.py", line 307, in sudo
     root_helper=self.root_helper)
   File "/usr/lib/python2.7/site-packages/oslo_concurrency/processutils.py", line 419, in execute
     cmd=sanitized_cmd)
 ProcessExecutionError: Unexpected error while running command.
 Command: lsblk -nlfoFSTYPE /e6c313a8-4e84-4b43-bca1-976ac1159825
 Exit code: 1
 Stdout: u''
 Stderr: u'lsblk: dm-2: failed to get device path\n'

This seems to happen only on slow connections to the storage, and a
couple of retries with waits seem to resolve the issue.

This patch adds this exit code to the errors we are allowed to retry.