Reorder cinderlib import
Python 2 has a bug on the builtin getaddrinfo method that incorrectly
fails when receiving a unicode string as the `port` parameter.  It only
accepts `str` and `int` values.

Cinderlib has a workaround for this, but for it to work cinderlib needs
to be imported before envetlet.

Reorder the cinderlib import to be before eventlet, which is something
that should have already been in that order in the first place due to
the alphabetical order.