Fix issues receiving duplicated RPC calls
We didn't handle concurrent duplicated RPC calls appropriately.

Now we track within the running service the requests for a specific
volume and raise ABORTED status code if we are already performing an
operation on that volume.

For the create volume call we use the name to detect simultaneous calls
instead the volume id, since this doesn't exist yet.