Add CRD persistence support
Only cinderlib's default persistence plugins were supported, which meant
that for the CSI use case we could only use the database plugin.

We are adding now our own CRD persistent metadata plugin.  Thanks to
this plugin in Kubernetes we won't need to be setup any database.

This enables deploying multiple Node services.

This is a rough first implementation using the official Kubernetes
client, which in our case may be more trouble than it's worth, as it
brings more dependencies and has some bugs we have to work around.