Add Kubernetes demo
This patch adds a Kubernetes demo that, with a single script, deploys a
Kubernetes cluster with a master and 2 additional nodes.  Configuring
the CSI driver with an LVM backend.

The demo also provides a step by step guide on how to test things and
how to change the default backend so we can try a different one.