Fix #22: K8s example - Second attach deletes data
Kubernetes example in the repository deploys CSI nodes in such a way
that a second attachment of the PV will format the volume again, losing
the data.

This happens because the lsblk command doesn't return the filesystem of
the block device on the second attach, so the system thinks it's empty
and formats it.

lsblk uses udev data to report the filesystem type, so we need to mount
/dev/udev for it to be able to report the filesystem type.

This patch mounts /run/udev with host to container propagation.