CRD plugin: Don't access backend name on the field
When the CRD persistence metadata plugin stores a volume it adds a label
for the backend name.  This lavel is retrieved using the name of the
OVO field in cinderlib that contains the name.

This patch stops using the field directly and uses the backend attribute
in the volume to access it's id (backend name) instead.

This means that we won't be affected when the backend name changes in
cinderlib to be stored as part of the host field like Cinder normally
does.