Fix volume's backend_name in CRD
In commit 5bd4dffe9107bc906d14a45cd819d9a659c19047 we modified the
getattr used to set the backend_name on the CRD labels, and we made it
so that it would only work when we had the full backend configured.

This patch fixes this and properly stores the backend_name regardless of
which type of service is updating it.