Fix CI requirements
On commit 6f8212c0c887491f2c10e4bf0197ab487f99c8f3 we removed the
cinderlib library from the requirements as they will be coming from the
Cinder respository.

This broke the CI that was depending on it.

Add the cinderlib requirement to the CI requirements.txt file to solve
this issue.