Simplify backend configuration
This patch tries to simplify the backend configuration by creating key
and value aliases on the X_CSI_BACKEND_CONFIG parameter.

Key aliases are:

- driver -> volume_driver
- multipath -> use_multipath_for_image_xfer
- name -> volume_backend_name

Value aliases only exist for the driver/volume_driver parameter, and
they are dynamically generated based on current Cinder installation.

So instead of having:
  "cinder.volume.drivers.lvm.LVMVolumeDriver"

We can now just use:
  "LVMVolume"

Which means that our configuration could change from:

  "volume_backend_name":"lvm",
  "volume_driver":"cinder.volume.drivers.lvm.LVMVolumeDriver",
  "use_multipath_for_image_xfer":false

To something more user friendly:

  "name":"lvm",
  "driver":"LVMVolume",
  "multipath":false

Since it can be hard for users to know which drivers are available on a
system we have included a CLI interface to get this listing through the
script "ember-list-drivers".

Closes: #62