Add rbd-nbd package to container
Now that cinderlib supports rbd-nbd to attach RBD volumes we should
include it in our containers, as it'd be the preferred connection tool
for RBD volumes to avoid feature support mismatch between the kernel
module and the Ceph cluster.

The Dockerfile includes it via the base cinderlib container image, and
this patch adds it to Dockerfile-master.

Closes: #63