Don't return volume size
We are returning a false value for Snapshots, as we are returning the
volume size, which in general will be greater than the actual snapshot
size.

Since the CSI spec allows us to return an undefined value we should do
so, since we can't tell the real size via cinderlib.

Fix: #60