Improve GRPC call log
On info level current code only logs the name of the method being
called, which is not enough to trace things.

Now we also log the ID (name, volume_id, or snapshot_id) of the resource
on the reception of the call, and the volume or snapshot_id on the
response so we can trace something at info level.