Support deleting volume with snapshots
This patch adds a soft-delete mechanism for volumes so they can be
"deleted" even when they have snapshots.

The expected behavior of deleting a volume with snapshots is not well
defined in the CSI spec, as stated in issue
https://github.com/container-storage-interface/spec/issues/346 but until
this is clarified we will assume that PR #347 will be accepted and we
must behave like this.

This is also how the csi-sanity test suite seems to expect us to behave.

Close: #70