Fix iSCSI issue on multi Ember deployment
We were storing each nodes connector information on a key defined only
by the node id, which is fine as long as all nodes share the Ember-CSI
deployments share the host network and the mount the /etc/iscsi
directory.

The problem comes when one of the Ember-CSI plugins don't use iSCSI, and
therefore it doesn't mount the /etc/iscsi directory, because in that
case we could be overwriting the initiator name of another plugin that
needs the host's initiator name.

An example:

- We deploy LVM iSCSI Ember-CSI and it works fine because the nodes are
sharing the host's initiator name and have stored it as a Key Value
using the node name as the key.

- We now deploy RBD Ember-CSI, which doesn't share the initiator name
with the host, so it stores the container's initiator name in the same
Key Value as the iSCSI plugin.

- Now when the iSCSI plugin tries to attach a volume, it will see the
initiator name stored by the RBD plugin and will therefore fail to
connect due to a credentials issue.

To fix this we now use the plugin name and the node id to form the key,
instead of just using the node id: PLUGIN_NAME + . + NODE_ID