Add support for readonly mode
In theory the volumes should be published by the controller as readonly,
but since we don't support that, we'll add the readonly mode by binding
on the NodePublishMethod with the ro option.

In terms of user experience this will be the same thing, since
containers won't be able to write to the volume or the mount.