Fix #92: Incorrect plugin name in CSI v1
CSI 0.x required reverse domain notation, but CSI
1.0 dropped the "reverse" part of the requirement. The expectation is that
the name will be normal (forward) domain notation.

So for CSI 1.0, instead of `io.ember-csi.[x]` the driver name should be
`[x].ember-csi.io`

This patch ensures that the right notation is used, depending on the CSI
version being used.

We also change the use of the "plugin_name" key in X_CSI_EMBER_CONFIG to
only refer to the the `[x]` part of the name, regardless of the notation
being used: `[x].ember-csi.io` or `io.ember-csi.[x]`.