Add subdomain from name to logs
When we have multiple Ember-CSI containers running on a host it is hard
to figure out what log entries belong to which plugin when looking at
the journaling in the host.

We are now logging the subdomain part of the name reported by the plugin
or default if there is no plugin name provided.

To reference this name when log formatting we must use "project_name".

Assuming we are using forward name because we are using CSI v1:

   plugin_name        CSI reported name   logged name
   -----------        -----------------   -----------
   rbd.ember-csi.io   rbd.ember-csi.io    rbd
   rbd                rbd.ember-csi.io    rbd
                      ember-csi.io        default