Fix #100 - Shared request id
All concurrent requests are sharing the same request ID, so when we
receive a new request it overwrites the request ID of the previous
thread and they all start reporting the same ID.

Timeline:

- Request #1 comes in --> Start logging as #1
- Request #2 comes in --> Start logging as #2
- Request #1 logs as request #2
- Request #2 logs as request #2
- Request #3 comes in --> Start logging as #3
- Request #1 logs as request #3
- Request #2 logs as request #3

This is caused by not monkey patching the threading library soon enough
and the thread storage is actually using the storage of the native
thread instead of locations for each greenthread.

There is only 1 case where the code will fail with this change, and that
is if the `hostname` is not in /etc/hosts as the greenthread DNS library
does not check /etc/hostname, only /etc/hosts.

The solution is to either add the hostname to /etc/hosts or set
environmental variable `EVENTLET_NO_GREENDNS=yes`.  But we don't
recommend the environmental variable, since this may also affect our
drivers if they are using hostnames in their configurations.