Fix docker hub build hook
The build process broke when we introduced multi-stage builds, and we
didn't detect it in the automatic build because the last command
executed in the script was successful.

This patch fixes the script and makes sure any error will abort the
execution of the code.

We change how we build containers, instead of doing a multi stage build
we always install using RDO RPM packages, from Trunk or a specific
release.