Fix #127: Fix NodeGetVolumeStats
The contents of /proc/self/mountinfo seem to be different when we are
running containerized and when we are not.

With current code, calls to NodeGetVolumeStats within a container with a
valid path will always fail with "Path does not match with requested
volume" when running.

This patch changes the column where we look for the mounted path based
on whether we are running inside a container or not.