Change csi-sanity for CSI spec v1
There is a newer version of the csi-sanity test suite that comes with
additional tests.  This patch updates the binary in the repository to
the latest version.

Latest tagged release of csi-sanity is making incorrect assumptions on
how pagination should work [1], so we cannot use that version until the
fix [2] has been released.  We use our own compiled version instead.

[1]: https://github.com/kubernetes-csi/csi-test/issues/204
[2]: https://github.com/kubernetes-csi/csi-test/pull/205