Close #108: Finer grained `vendor_version`
This patch makes Ember-CSI report a finer grained `vendor_version` in
the `GetPluginCapabilities` response.

Reported version will depend on whether it's a release or a build from
master.

For build from master it will be reporting the Ember-CSI version, the
number of commits since the last release, the short hash, and the date
and time of the build as DDMMYYYYhhmmssnnnnnnnnn. For example:

  0.9.0-12-gbbce4e1+03092019170848330555972

For releases it will only append the date and time:

  0.9.0+03092019170848330555972

All versions reported conform to semver as per
https://semver.org/#spec-item-10

For master builds it will also set this version as the Ember-CSI Python
package and in ember_csi.__version__.