Close #122: Support disabling features
This patch adds the possiblity of disabling some of the optional CSI
features that were previously enabled by default.

Reasons to disable features are:

- Feature is not efficient (ie: cloning doing dd)
- Feature requires a license in the backend (ie: snapshot for NetApp)

Now we can disable them passing a list of elements on the "disabled" key
in the X_CSI_EMBER_CONFIG environmental variable dictionary.

We can disable 2 features:

- Snapshots: With the "snapshot" string
- Cloning: With the "clone" string

Disabling snapshots will disable snapshot listing as well and will
return unimplemented if the CO has a bug and makes the request.

When we disable cloning the error we'll get if we try to use it is
INVALID_ARGUMENT, same thing will happen if we disable snapshots and try
to create a volume from a snapshot.