Fix #140: Volume from snap doesn't extend volume
When we create a volume from a snapshot and request that the new volume
has a greater size than the original snapshot the volume ends up with
the same size as the snapshot instead of the requested size.

This patch passes the requested volume size to the snapshot's
create_volume call, as we should have done all along.