Fix cinderlib LVM extend volume
Cinderlib has a bug (#1852629) such that when we try to extend an
attached LVM volume after retrieving the volume metadata from the
persistent storage we get a TargetUpdateFailed exception cause by
exception AttributeError: 'DB' object has no attribute
'volume_attachment_get_all_by_volume_id'.

This is caused by the LVM driver checking volume_attachment which is not
populated, which triggers a lazy loading on the OVO which calls a method
in the DB layer that is not available in our fake DB instance.

Until upstream merges https://review.opendev.org/694524 and the RPM is
available we will just patch the code with the contents of the upsteram
patch (excluding the tests partis).