Support RBD volume extend
The RBD OS-Brick code is missing the extend_volume method because there
is no need to call it for RBD volumes, since the RBD protocol
automatically informs the host and updates the size.

Cinderlib is also lacking the the local extend in the connector.

In this patch, until we have RPMs available, we import code changes from
upstream review https://review.opendev.org/#/c/690713 and add it to the
cinderlib changes we make on installation.