Fix RBD in k8s 1.13 example
The RBD backend isn't working anymore because the Ceph Demo container
now creates a different ceph.conf file, and appending the features line
at the end of the file doesn't add it to the global section anymore.

We fix this by ensuring we always add the features configuration option
as the first option in the global section.