Example k8s 1.13 support restart node container
Restarting the Node csi-driver container when it has attached volumes
would break the CSI plugin, since it could no longer detach the volumes,
because it loses the private references for the volumes.

The Ansible playbooks had a comment that production environments should
be doing things differently.

This patch replaces the comment with the jobs to actually do the right
mounts.