Fix #148: NodeGetVolumeStats returning NOT_FOUND
Back when we fixed this same issue (bug #127) the code made the
incorrect assumption of thinking that the position of the mount source
depends on whether the service is running containerized or not, when it
actually depends on the number of optional fields.

This patch makes sure we properly parse the mountinfo file to account
for the optional fields as well as accept staging and publish paths as
the volume_path parameter in the call.