Fix #153: Example k8s v1.16 RegisterPlugin error
We are continuously seeing the following errors in the Kubernetes 1.16
example:

  Error: "RegisterPlugin error -- failed to get plugin info using RPC
  GetInfo at socket /var/lib/kubelet/plugins/ember-csi.io/csi.sock, err:
  rpc error: code = Unimplemented desc = Method not found!"

This is because k8s 1.15 introduced a change in the plugin-manager that
keeps retrying till all the plugins (socket files under
/var/lib/kubelet/{plugins,plugin_registry}) registrations succeed.  That
didn't use to be the case in previous releases, where we would only see
this error once.

This patch changes the directory for the Ember-CSI locks in the example
and also removes the cluster-driver-registrar on the controller pods
since they are no longer necessary (they have been replace by us
creating manually the CSIDriver object).