Fix #155: Attach volume from source without FS
When we create a mount type of volume from another volume or snapshot
we may find ourselves in a situation where the new volume cannot be
attached as the mount operation fails.

This is because we can create a volume from a source that has created
the filesystem on the device but has not flushed it to the backend yet,
so when we create another volume from it we end up with a volume that
has been marked with a specific filesystem but its contents don't really
have the filesystem yet.

So the code will not format it thinking that it already has the
filesystem and fail when it tries to mount it.

With this patch we change the way we detect whether to format a mount
type of volume or not.  We now always do a check of the volume's
filesystem type using lsblk to confirm that the volume has the expected
filesystem.

We also present a meaningful warning if we detect this condition where
the metadata of the volume and the actual filesystem don't match.