Fix issue with required size < 1BGi
When we received a requests where the "required" size was <1GBi but
there was no "limit" we would complain that the minimum size we can
create was 1GBi.

This patch fixes the way we detect unsupported sizes in the
"_calculate_size" method as well as return a proper "max_size".

Fix #158