Add py2 and py3 csi_pb2 files
Now that we support Python 2 and 3 it would be nice to have independent
csi_pb2.py and csi_pb2_grpc.py files for each Python version.

This patch adds the necessary mechanism to have this and support the
installation without having the unnecessary files included.

While developing we can manually change the files using:

  python setup.py csi_proto