Fix cinderlib rbd attach
Attaching an RBD volume fails when we run the code as root if we are
using Python 3.

We see the following error:

   conn.attach() File
   "/usr/lib/python3.6/site-packages/cinderlib/objects.py", line 842, in
   attach reason=error_msg) cinder.exception.DeviceUnavailable: The
   device in the path None is unavailable: Unable to access the backend
   storage via path /dev/rbd0.

In this patch, until we have RPMs available, we import code changes from
upstream review https://review.opendev.org/#/c/720553/ and add it to the
cinderlib changes we make on installation.