Centos 8 Dockerfile
Now that we support Python 3 we generate CentOS 8 containers as well as
CentOS 7.

We leave the default containers to be CentOS 7 with Python 2, which are
also tagged with the postfix of 7.  For example we'll have
ember-csi:master and ember-csi:master7 tagging the same container.

The CentOS 8 images are postfixed with an 8, ie: ember-csi:master8.