Improve apply-patches
The default installation of patches/apply-patches uses a hardcoded
value, which will be incorrect in most cases, so we will query python
instead.

The messages during installation are misleading:

  Patching /usr/lib/python3.6/site-packages with cinderlib-bug-1856556.patch
  Patch already applied, skipping
  Patching /usr/lib/python3.6/site-packages with cinderlib-bug-1873312.patch
  1 out of 1 hunk FAILED
  1 out of 1 hunk FAILED

So it's hard to tell if the FAILED messages should have been a failure
or not.

So we update the script to have a clearer output:

  Checking if cinderlib-bug-1856556.patch is already present in /home/geguileo/code/reuse-cinder-drivers/ember/.venv3/lib/python3.7/site-packages
      Patch already applied, skipping
  Checking if cinderlib-bug-1873312.patch is already present in /home/geguileo/code/reuse-cinder-drivers/ember/.venv3/lib/python3.7/site-packages
      Patch not present, applying it