Flush buffers on node unpublish
According to k8s e2e csi tests they expect the data to be flushed after
the unpublish, but in our case we weren't flushing filesystems until the
volume was unstaged.

With this patch we now force the sync of the filesystem and block volume
on the unstage phase.

With this change the e2e tests should no longer fail on tests like
volume cloning.