Remove centos 7 container
The cinderlib and cinder projects now only support Python 3, and we
updated the Dockerfile to properly use them for master.

Unfortunately just a couple of days later upstream changed the
constraints in OpenStack, bringing in a new version of the cryptography
library which required Rust, setuptools-rust, and some other changes.

We could fix it now, but it will keep happening over and over again.

Now that we no longer need to keep the iscsid binaries in sync between
ember and the system, because we are using nsenter, we can have a single
container for all deployments.

After this point we will no longer be building Centos 7 based images.