Fix idempotency issue on snapshot creation
This patch fixes our behavior when creating snapshots, previously if we
found that something had gone wrong and we had multiple snapshots with
the same name (which should be unique) we ignored it and created a new
one.

This patch changes this behavior and considers this situation to be an
abnormal state and rejects the creation, leaving it for an administrator
to fix.