Fix pagination
For pagination we use tokens based on the creation datetime of the
resource, but we are currently only using second resolution, which leads
to us doing pagination wrong if multiple resources have been created
within the same second.

This patch fixes this by using the microseconds from the datetime to
increase the resolution.

Currently cinderlib has a similar bug [1] that removes this microsecond
information when converting to JSON, but once that bug is fixed
Ember-CSI will be able to use the additional resolution.

[1]: https://bugs.launchpad.net/cinderlib/+bug/1933964