Fix XFS volumes from data source
As mentioned in a CSI issue [1], XFS does not allow to mount two volumes
that have the same UUID on the same machine.

This is problematic when using a cloned volume or a restored snapshot -
the original volume and the new volume cannot be mounted on the same
compute node.

This patch fixes this issue by mounting XFS volumes with the "nouuid"
option, since regenerating the UUID requires mounting the volume on the
controller, and now the latest e2e tests [1] pass.

[1]: https://github.com/container-storage-interface/spec/issues/482
[2]: https://github.com/kubernetes/kubernetes/pull/102538